+++ drift/lib/src/runtime/devtools/service_extension.dart
@@ -5,6 +5,7 @@ import 'dart:developer';
 import 'package:drift/drift.dart';
 import 'package:drift/src/remote/protocol.dart';
 import 'package:drift/src/runtime/executor/transactions.dart';
+import 'package:meta/meta.dart';
 
 import '../api/runtime_api.dart';
 import 'devtools.dart';
@@ -63,7 +64,7 @@ class DriftServiceExtension {
 
         return _protocol.encodePayload(result);
       case 'collect-expected-schema':
-        final executor = _CollectCreateStatements();
+        final executor = CollectCreateStatements(SqlDialect.sqlite);
         await tracked.database.runConnectionZoned(
             BeforeOpenRunner(tracked.database, executor), () async {
           // ignore: invalid_use_of_protected_member, invalid_use_of_visible_for_testing_member
@@ -109,8 +110,13 @@ class DriftServiceExtension {
   static const _protocol = DriftProtocol();
 }
 
-class _CollectCreateStatements extends QueryExecutor {
+@internal
+final class CollectCreateStatements extends QueryExecutor {
   final List<String> statements = [];
+  @override
+  final SqlDialect dialect;
+
+  CollectCreateStatements(this.dialect);
 
   @override
   QueryExecutor beginExclusive() {
@@ -121,9 +127,6 @@ class _CollectCreateStatements extends QueryExecutor {
   TransactionExecutor beginTransaction() {
     throw UnimplementedError();
   }
-
-  @override
-  SqlDialect get dialect => SqlDialect.sqlite;
 
   @override
   Future<bool> ensureOpen(QueryExecutorUser user) {